refactor(ingress): tighten correlation diagnostics
